+++ tensorflow_decision_forests/keras/wrappers_pre_generated.py
@@ -31,6 +31,7 @@ import tensorflow as tf
 
 from tensorflow_decision_forests.keras import core
 from yggdrasil_decision_forests.model import abstract_model_pb2  # pylint: disable=unused-import
+from yggdrasil_decision_forests.learner import abstract_learner_pb2
 
 TaskType = "abstract_model_pb2.Task"  # pylint: disable=invalid-name
 AdvancedArguments = core.AdvancedArguments
@@ -104,6 +105,11 @@ class CartModel(core.CoreModel):
       efficiency. If specified, `num_threads` field of the
       `advanced_arguments.yggdrasil_deployment_config` has priority.
     name: The name of the model.
+    max_vocab_count: Default maximum size of the vocabulary for CATEGORICAL and
+      CATEGORICAL_SET features stored as strings. If more unique values exist,
+      only the most frequent values are kept, and the remaining values are
+      considered as out-of-vocabulary. The value `max_vocab_count` defined in a
+      `FeatureUsage` (if any) takes precedence.
     allow_na_conditions: If true, the tree training evaluates conditions of the
       type `X is NA` i.e. `X is missing`. Default: False.
     categorical_algorithm: How to learn splits on categorical attributes.
@@ -153,11 +159,21 @@ class CartModel(core.CoreModel):
       (i.e. a node can be derived only if it contains more than `min_examples`
       examples). If false, there can be nodes with less than
       `min_examples` training examples. Default: True.
+    keep_non_leaf_label_distribution: Whether to keep the node value (i.e. the
+      distribution of the labels of the training examples) of non-leaf nodes.
+      This information is not used during serving, however it can be used for
+      model interpretation as well as hyper parameter tuning. In the worst
+      case, this can account for half of the model size. Default: True.
     max_depth: Maximum depth of the tree. `max_depth=1` means that all trees
       will be roots. Negative values are ignored. Default: 16.
     max_num_nodes: Maximum number of nodes in the tree. Set to -1 to disable
       this limit. Only available for `growing_strategy=BEST_FIRST_GLOBAL`.
       Default: None.
+    maximum_model_size_in_memory_in_bytes: Limit the size of the model when
+      stored in ram. Different algorithms can enforce this limit differently.
+      Note that when models are compiled into an inference, the size of the
+      inference engine is generally much smaller than the original model.
+      Default: -1.0.
     maximum_training_duration_seconds: Maximum training duration of the model
       expressed in seconds. Each learning algorithm is free to use this
       parameter at it sees fit. Enabling maximum training duration makes the
@@ -188,6 +204,8 @@ class CartModel(core.CoreModel):
       number_of_input_features x num_candidate_attributes_ratio`. The possible
       values are between ]0, and 1] as well as -1. If not set or equal to -1,
       the `num_candidate_attributes` is used. Default: -1.0.
+    random_seed: Random seed for the training of the model. Learners are
+      expected to be deterministic by the random seed. Default: 123456.
     sorting_strategy: How are sorted the numerical features in order to find the
       splits
       - PRESORT: The features are pre-sorted at the start of the training. This
@@ -237,6 +255,7 @@ class CartModel(core.CoreModel):
                advanced_arguments: Optional[AdvancedArguments] = None,
                num_threads: Optional[int] = 6,
                name: Optional[str] = None,
+               max_vocab_count: Optional[int] = 2000,
                allow_na_conditions: Optional[bool] = False,
                categorical_algorithm: Optional[str] = "CART",
                categorical_set_split_greedy_sampling: Optional[float] = 0.1,
@@ -244,13 +263,16 @@ class CartModel(core.CoreModel):
                categorical_set_split_min_item_frequency: Optional[int] = 1,
                growing_strategy: Optional[str] = "LOCAL",
                in_split_min_examples_check: Optional[bool] = True,
+               keep_non_leaf_label_distribution: Optional[bool] = True,
                max_depth: Optional[int] = 16,
                max_num_nodes: Optional[int] = None,
+               maximum_model_size_in_memory_in_bytes: Optional[float] = -1.0,
                maximum_training_duration_seconds: Optional[float] = -1.0,
                min_examples: Optional[int] = 5,
                missing_value_policy: Optional[str] = "GLOBAL_IMPUTATION",
                num_candidate_attributes: Optional[int] = 0,
                num_candidate_attributes_ratio: Optional[float] = -1.0,
+               random_seed: Optional[int] = 123456,
                sorting_strategy: Optional[str] = "PRESORT",
                sparse_oblique_normalization: Optional[str] = None,
                sparse_oblique_num_projections_exponent: Optional[float] = None,
@@ -274,10 +296,14 @@ class CartModel(core.CoreModel):
             growing_strategy,
         "in_split_min_examples_check":
             in_split_min_examples_check,
+        "keep_non_leaf_label_distribution":
+            keep_non_leaf_label_distribution,
         "max_depth":
             max_depth,
         "max_num_nodes":
             max_num_nodes,
+        "maximum_model_size_in_memory_in_bytes":
+            maximum_model_size_in_memory_in_bytes,
         "maximum_training_duration_seconds":
             maximum_training_duration_seconds,
         "min_examples":
@@ -288,6 +314,8 @@ class CartModel(core.CoreModel):
             num_candidate_attributes,
         "num_candidate_attributes_ratio":
             num_candidate_attributes_ratio,
+        "random_seed":
+            random_seed,
         "sorting_strategy":
             sorting_strategy,
         "sparse_oblique_normalization":
@@ -320,11 +348,198 @@ class CartModel(core.CoreModel):
         verbose=verbose,
         advanced_arguments=advanced_arguments,
         num_threads=num_threads,
-        name=name)
+        name=name,
+        max_vocab_count=max_vocab_count)
 
   @staticmethod
   def predefined_hyperparameters() -> List[core.HyperParameterTemplate]:
     return []
+
+  @staticmethod
+  def capabilities() -> abstract_learner_pb2.LearnerCapabilities:
+    return abstract_learner_pb2.LearnerCapabilities(
+        support_partial_cache_dataset_format=False)
+
+
+class DistributedGradientBoostedTreesModel(core.CoreModel):
+  r"""Distributed Gradient Boosted Trees learning algorithm.
+
+  Exact distributed version of the Gradient Boosted Tree learning algorithm. See
+  the documentation of the non-distributed Gradient Boosted Tree learning
+  algorithm for an introduction to GBTs.
+
+  Usage example:
+
+  ```python
+  import tensorflow_decision_forests as tfdf
+  import pandas as pd
+
+  dataset = pd.read_csv("project/dataset.csv")
+  tf_dataset = tfdf.keras.pd_dataframe_to_tf_dataset(dataset, label="my_label")
+
+  model = tfdf.keras.DistributedGradientBoostedTreesModel()
+  model.fit(tf_dataset)
+
+  print(model.summary())
+  ```
+
+  Attributes:
+    task: Task to solve (e.g. Task.CLASSIFICATION, Task.REGRESSION,
+      Task.RANKING).
+    features: Specify the list and semantic of the input features of the model.
+      If not specified, all the available features will be used. If specified
+      and if `exclude_non_specified_features=True`, only the features in
+      `features` will be used by the model. If "preprocessing" is used,
+      `features` corresponds to the output of the preprocessing. In this case,
+      it is recommended for the preprocessing to return a dictionary of tensors.
+    exclude_non_specified_features: If true, only use the features specified in
+      `features`.
+    preprocessing: Functional keras model or @tf.function to apply on the input
+      feature before the model to train. This preprocessing model can consume
+      and return tensors, list of tensors or dictionary of tensors. If
+      specified, the model only "sees" the output of the preprocessing (and not
+      the raw input). Can be used to prepare the features or to stack multiple
+      models on top of each other. Unlike preprocessing done in the tf.dataset,
+      the operation in "preprocessing" are serialized with the model.
+    postprocessing: Like "preprocessing" but applied on the model output.
+    ranking_group: Only for `task=Task.RANKING`. Name of a tf.string feature
+      that identifies queries in a query/document ranking task. The ranking
+      group is not added automatically for the set of features if
+      `exclude_non_specified_features=false`.
+    temp_directory: Temporary directory used to store the model Assets after the
+      training, and possibly as a work directory during the training. This
+      temporary directory is necessary for the model to be exported after
+      training e.g. `model.save(path)`. If not specified, `temp_directory` is
+      set to a temporary directory using `tempfile.TemporaryDirectory`. This
+      directory is deleted when the model python object is garbage-collected.
+    verbose: If true, displays information about the training.
+    hyperparameter_template: Override the default value of the hyper-parameters.
+      If None (default) the default parameters of the library are used. If set,
+      `default_hyperparameter_template` refers to one of the following
+      preconfigured hyper-parameter sets. Those sets outperforms the default
+      hyper-parameters (either generally or in specific scenarios). You can omit
+      the version (e.g. remove "@v5") to use the last version of the template.
+      In this case, the hyper-parameter can change in between releases (not
+      recommended for training in production).
+    advanced_arguments: Advanced control of the model that most users won't need
+      to use. See `AdvancedArguments` for details.
+    num_threads: Number of threads used to train the model. Different learning
+      algorithms use multi-threading differently and with different degree of
+      efficiency. If specified, `num_threads` field of the
+      `advanced_arguments.yggdrasil_deployment_config` has priority.
+    name: The name of the model.
+    max_vocab_count: Default maximum size of the vocabulary for CATEGORICAL and
+      CATEGORICAL_SET features stored as strings. If more unique values exist,
+      only the most frequent values are kept, and the remaining values are
+      considered as out-of-vocabulary. The value `max_vocab_count` defined in a
+      `FeatureUsage` (if any) takes precedence.
+    apply_link_function: If true, applies the link function (a.k.a. activation
+      function), if any, before returning the model prediction. If false,
+      returns the pre-link function model output. For example, in the case of
+      binary classification, the pre-link function
+      output is a logic while the post-link function is a probability. Default:
+        True.
+    max_depth: Maximum depth of the tree. `max_depth=1` means that all trees
+      will be roots. Negative values are ignored. Default: 6.
+    maximum_model_size_in_memory_in_bytes: Limit the size of the model when
+      stored in ram. Different algorithms can enforce this limit differently.
+      Note that when models are compiled into an inference, the size of the
+      inference engine is generally much smaller than the original model.
+      Default: -1.0.
+    maximum_training_duration_seconds: Maximum training duration of the model
+      expressed in seconds. Each learning algorithm is free to use this
+      parameter at it sees fit. Enabling maximum training duration makes the
+      model training non-deterministic. Default: -1.0.
+    min_examples: Minimum number of examples in a node. Default: 5.
+    num_trees: Maximum number of decision trees. The effective number of
+      trained tree can be smaller if early stopping is enabled. Default: 300.
+    random_seed: Random seed for the training of the model. Learners are
+      expected to be deterministic by the random seed. Default: 123456.
+    shrinkage: Coefficient applied to each tree prediction. A small value (0.02)
+      tends to give more accurate results (assuming enough trees are trained),
+      but results in larger models. Analogous to neural network
+      learning rate. Default: 0.1.
+    use_hessian_gain: Use true, uses a formulation of split gain with a hessian
+      term i.e. optimizes the splits to minimize the variance of "gradient /
+      hessian. Available for all losses except regression. Default: False.
+  """
+
+  @core._list_explicit_arguments
+  def __init__(self,
+               task: Optional[TaskType] = core.Task.CLASSIFICATION,
+               features: Optional[List[core.FeatureUsage]] = None,
+               exclude_non_specified_features: Optional[bool] = False,
+               preprocessing: Optional["tf.keras.models.Functional"] = None,
+               postprocessing: Optional["tf.keras.models.Functional"] = None,
+               ranking_group: Optional[str] = None,
+               temp_directory: Optional[str] = None,
+               verbose: Optional[bool] = True,
+               hyperparameter_template: Optional[str] = None,
+               advanced_arguments: Optional[AdvancedArguments] = None,
+               num_threads: Optional[int] = 6,
+               name: Optional[str] = None,
+               max_vocab_count: Optional[int] = 2000,
+               apply_link_function: Optional[bool] = True,
+               max_depth: Optional[int] = 6,
+               maximum_model_size_in_memory_in_bytes: Optional[float] = -1.0,
+               maximum_training_duration_seconds: Optional[float] = -1.0,
+               min_examples: Optional[int] = 5,
+               num_trees: Optional[int] = 300,
+               random_seed: Optional[int] = 123456,
+               shrinkage: Optional[float] = 0.1,
+               use_hessian_gain: Optional[bool] = False,
+               explicit_args: Optional[Set[str]] = None):
+
+    learner_params = {
+        "apply_link_function":
+            apply_link_function,
+        "max_depth":
+            max_depth,
+        "maximum_model_size_in_memory_in_bytes":
+            maximum_model_size_in_memory_in_bytes,
+        "maximum_training_duration_seconds":
+            maximum_training_duration_seconds,
+        "min_examples":
+            min_examples,
+        "num_trees":
+            num_trees,
+        "random_seed":
+            random_seed,
+        "shrinkage":
+            shrinkage,
+        "use_hessian_gain":
+            use_hessian_gain,
+    }
+
+    if hyperparameter_template is not None:
+      learner_params = core._apply_hp_template(
+          learner_params, hyperparameter_template,
+          self.predefined_hyperparameters(), explicit_args)
+
+    super(DistributedGradientBoostedTreesModel, self).__init__(
+        task=task,
+        learner="DISTRIBUTED_GRADIENT_BOOSTED_TREES",
+        learner_params=learner_params,
+        features=features,
+        exclude_non_specified_features=exclude_non_specified_features,
+        preprocessing=preprocessing,
+        postprocessing=postprocessing,
+        ranking_group=ranking_group,
+        temp_directory=temp_directory,
+        verbose=verbose,
+        advanced_arguments=advanced_arguments,
+        num_threads=num_threads,
+        name=name,
+        max_vocab_count=max_vocab_count)
+
+  @staticmethod
+  def predefined_hyperparameters() -> List[core.HyperParameterTemplate]:
+    return []
+
+  @staticmethod
+  def capabilities() -> abstract_learner_pb2.LearnerCapabilities:
+    return abstract_learner_pb2.LearnerCapabilities(
+        support_partial_cache_dataset_format=True)
 
 
 class GradientBoostedTreesModel(core.CoreModel):
@@ -370,9 +585,9 @@ class GradientBoostedTreesModel(core.CoreModel):
       models on top of each other. Unlike preprocessing done in the tf.dataset,
       the operation in "preprocessing" are serialized with the model.
     postprocessing: Like "preprocessing" but applied on the model output.
-    ranking_group: Only for `task=Task.RANKING`. Name of a tf.string feature that
-      identifies queries in a query/document ranking task. The ranking group
-      is not added automatically for the set of features if
+    ranking_group: Only for `task=Task.RANKING`. Name of a tf.string feature
+      that identifies queries in a query/document ranking task. The ranking
+      group is not added automatically for the set of features if
       `exclude_non_specified_features=false`.
     temp_directory: Temporary directory used to store the model Assets after the
       training, and possibly as a work directory during the training. This
@@ -385,19 +600,18 @@ class GradientBoostedTreesModel(core.CoreModel):
       If None (default) the default parameters of the library are used. If set,
       `default_hyperparameter_template` refers to one of the following
       preconfigured hyper-parameter sets. Those sets outperforms the default
-      hyper-parameters (either generally or in specific scenarios).
-      You can omit the version (e.g. remove "@v5") to use the last version of
-      the template. In this case, the hyper-parameter can change in between
-      releases (not recommended for training in production).
+      hyper-parameters (either generally or in specific scenarios). You can omit
+      the version (e.g. remove "@v5") to use the last version of the template.
+      In this case, the hyper-parameter can change in between releases (not
+      recommended for training in production).
       - better_default@v1: A configuration that is generally better than the
         default parameters without being more expensive. The parameters are:
-        growing_strategy="BEST_FIRST_GLOBAL".
+          growing_strategy="BEST_FIRST_GLOBAL".
       - benchmark_rank1@v1: Top ranking hyper-parameters on our benchmark
         slightly modified to run in reasonable time. The parameters are:
-        growing_strategy="BEST_FIRST_GLOBAL", categorical_algorithm="RANDOM",
-        split_axis="SPARSE_OBLIQUE", sparse_oblique_normalization="MIN_MAX",
-        sparse_oblique_num_projections_exponent=1.0.
-
+          growing_strategy="BEST_FIRST_GLOBAL", categorical_algorithm="RANDOM",
+          split_axis="SPARSE_OBLIQUE", sparse_oblique_normalization="MIN_MAX",
+          sparse_oblique_num_projections_exponent=1.0.
     advanced_arguments: Advanced control of the model that most users won't need
       to use. See `AdvancedArguments` for details.
     num_threads: Number of threads used to train the model. Different learning
@@ -405,6 +619,11 @@ class GradientBoostedTreesModel(core.CoreModel):
       efficiency. If specified, `num_threads` field of the
       `advanced_arguments.yggdrasil_deployment_config` has priority.
     name: The name of the model.
+    max_vocab_count: Default maximum size of the vocabulary for CATEGORICAL and
+      CATEGORICAL_SET features stored as strings. If more unique values exist,
+      only the most frequent values are kept, and the remaining values are
+      considered as out-of-vocabulary. The value `max_vocab_count` defined in a
+      `FeatureUsage` (if any) takes precedence.
     adapt_subsample_for_maximum_training_duration: Control how the maximum
       training duration (if set) is applied. If false, the training stop when
       the time is used. If true, the size of the sampled datasets used train
@@ -414,17 +633,16 @@ class GradientBoostedTreesModel(core.CoreModel):
       type `X is NA` i.e. `X is missing`. Default: False.
     apply_link_function: If true, applies the link function (a.k.a. activation
       function), if any, before returning the model prediction. If false,
-      returns the pre-link function model output.
-      For example, in the case of binary classification, the pre-link function
+      returns the pre-link function model output. For example, in the case of
+      binary classification, the pre-link function
       output is a logic while the post-link function is a probability. Default:
-      True.
+        True.
     categorical_algorithm: How to learn splits on categorical attributes.
       - `CART`: CART algorithm. Find categorical splits of the form "value \\in
         mask". The solution is exact for binary classification, regression and
         ranking. It is approximated for multi-class classification. This is a
-        good first algorithm to use. In case of overfitting (very small
-        dataset, large dictionary), the "random" algorithm is a good
-        alternative.
+        good first algorithm to use. In case of overfitting (very small dataset,
+        large dictionary), the "random" algorithm is a good alternative.
       - `ONE_HOT`: One-hot encoding. Find the optimal categorical split of the
         form "attribute == param". This method is similar (but more efficient)
         than converting converting each possible categorical value into a
@@ -445,11 +663,15 @@ class GradientBoostedTreesModel(core.CoreModel):
       available, the least frequent items are ignored. Changing this value is
       similar to change the "max_vocab_count" before loading the dataset, with
       the following exception: With `max_vocab_count`, all the remaining items
-      are grouped in a special Out-of-vocabulary item. With `max_num_items`,
+        are grouped in a special Out-of-vocabulary item. With `max_num_items`,
       this is not the case. Default: -1.
     categorical_set_split_min_item_frequency: For categorical set splits e.g.
       texts. Minimum number of occurrences of an item to be considered.
       Default: 1.
+    compute_permutation_variable_importance: If true, compute the permutation
+      variable importance of the model at the end of the training using the
+      validation dataset. Enabling this feature can increase the training time
+      significantly. Default: False.
     dart_dropout: Dropout rate applied when using the DART i.e. when
       forest_extraction=DART. Default: 0.01.
     early_stopping: Early stopping detects the overfitting of the model and
@@ -460,7 +682,7 @@ class GradientBoostedTreesModel(core.CoreModel):
         truncated to maximize the validation loss.
       - `LOSS_INCREASE`: Stop the training when the validation does not
         decrease for `early_stopping_num_trees_look_ahead` trees. Default:
-        "LOSS_INCREASE".
+          "LOSS_INCREASE".
     early_stopping_num_trees_look_ahead: Rolling number of trees used to detect
       validation loss increase and trigger early stopping. Default: 30.
     forest_extraction: How to construct the forest:
@@ -469,7 +691,7 @@ class GradientBoostedTreesModel(core.CoreModel):
         previous trees.
       - DART: For Dropout Additive Regression Trees. A modification of MART
         proposed in http://proceedings.mlr.press/v38/korlakaivinayak15.pdf.
-        Here, each tree tries to "correct" the mistakes of a random subset of
+          Here, each tree tries to "correct" the mistakes of a random subset of
         the previous trees. Default: "MART".
     goss_alpha: Alpha parameter for the GOSS (Gradient-based One-Side Sampling;
       "See LightGBM: A Highly Efficient Gradient Boosting Decision Tree")
@@ -481,17 +703,22 @@ class GradientBoostedTreesModel(core.CoreModel):
         words, as long as a node satisfy the splits "constraints (e.g. maximum
         depth, minimum number of observations), the node will be split. This is
         the "classical" way to grow decision trees.
-      - `BEST_FIRST_GLOBAL`: The node with the best loss reduction among all
-        the nodes of the tree is selected for splitting. This method is also
-        called "best first" or "leaf-wise growth". See "Best-first decision
+      - `BEST_FIRST_GLOBAL`: The node with the best loss reduction among all the
+        nodes of the tree is selected for splitting. This method is also called
+        "best first" or "leaf-wise growth". See "Best-first decision
         tree learning", Shi and "Additive logistic regression : A statistical
         view of boosting", Friedman for more details. Default: "LOCAL".
     in_split_min_examples_check: Whether to check the `min_examples` constraint
       in the split search (i.e. splits leading to one child having less than
-      `min_examples` examples are considered invalid) or before the split
-      search (i.e. a node can be derived only if it contains more than
-      `min_examples` examples). If false, there can be nodes with less than
+      `min_examples` examples are considered invalid) or before the split search
+      (i.e. a node can be derived only if it contains more than `min_examples`
+      examples). If false, there can be nodes with less than
       `min_examples` training examples. Default: True.
+    keep_non_leaf_label_distribution: Whether to keep the node value (i.e. the
+      distribution of the labels of the training examples) of non-leaf nodes.
+      This information is not used during serving, however it can be used for
+      model interpretation as well as hyper parameter tuning. In the worst
+      case, this can account for half of the model size. Default: True.
     l1_regularization: L1 regularization applied to the training loss. Impact
       the tree structures and lead values. Default: 0.0.
     l2_categorical_regularization: L2 regularization applied to the training
@@ -506,15 +733,20 @@ class GradientBoostedTreesModel(core.CoreModel):
     max_num_nodes: Maximum number of nodes in the tree. Set to -1 to disable
       this limit. Only available for `growing_strategy=BEST_FIRST_GLOBAL`.
       Default: None.
+    maximum_model_size_in_memory_in_bytes: Limit the size of the model when
+      stored in ram. Different algorithms can enforce this limit differently.
+      Note that when models are compiled into an inference, the size of the
+      inference engine is generally much smaller than the original model.
+      Default: -1.0.
     maximum_training_duration_seconds: Maximum training duration of the model
       expressed in seconds. Each learning algorithm is free to use this
       parameter at it sees fit. Enabling maximum training duration makes the
       model training non-deterministic. Default: -1.0.
     min_examples: Minimum number of examples in a node. Default: 5.
     missing_value_policy: Method used to handle missing attribute values.
-      - `GLOBAL_IMPUTATION`: Missing attribute values are imputed, with the
-        mean (in case of numerical attribute) or the most-frequent-item (in
-        case of categorical attribute) computed on the entire dataset (i.e. the
+      - `GLOBAL_IMPUTATION`: Missing attribute values are imputed, with the mean
+        (in case of numerical attribute) or the most-frequent-item (in case of
+        categorical attribute) computed on the entire dataset (i.e. the
         information contained in the data spec).
       - `LOCAL_IMPUTATION`: Missing attribute values are imputed with the mean
         (numerical attribute) or most-frequent-item (in the case of categorical
@@ -528,8 +760,8 @@ class GradientBoostedTreesModel(core.CoreModel):
       node. An attribute is valid if it has at least a valid split. If
       `num_candidate_attributes=0`, the value is set to the classical default
       value for Random Forest: `sqrt(number of input attributes)` in case of
-      classification and `number_of_input_attributes / 3` in case of
-      regression. If `num_candidate_attributes=-1`, all the attributes are
+        classification and `number_of_input_attributes / 3` in case of
+        regression. If `num_candidate_attributes=-1`, all the attributes are
       tested. Default: -1.
     num_candidate_attributes_ratio: Ratio of attributes tested at each node. If
       set, it is equivalent to `num_candidate_attributes =
@@ -538,6 +770,8 @@ class GradientBoostedTreesModel(core.CoreModel):
       the `num_candidate_attributes` is used. Default: -1.0.
     num_trees: Maximum number of decision trees. The effective number of
       trained tree can be smaller if early stopping is enabled. Default: 300.
+    random_seed: Random seed for the training of the model. Learners are
+      expected to be deterministic by the random seed. Default: 123456.
     sampling_method: Control the sampling of the datasets used to train
       individual trees.
       - NONE: No sampling is applied.
@@ -553,20 +787,20 @@ class GradientBoostedTreesModel(core.CoreModel):
       Boosting for Effective Learning to Rank; Lucchese et al;
       http://quickrank.isti.cnr.it/selective-data/selective-SIGIR2018.pdf)
       sampling method. Default: 0.01.
-    shrinkage: Coefficient applied to each tree prediction. A small value
-      (0.02) tends to give more accurate results (assuming enough trees are
-      trained), but results in larger models. Analogous to neural network
+    shrinkage: Coefficient applied to each tree prediction. A small value (0.02)
+      tends to give more accurate results (assuming enough trees are trained),
+      but results in larger models. Analogous to neural network
       learning rate. Default: 0.1.
-    sorting_strategy: How are sorted the numerical features in order to find
-      the splits
+    sorting_strategy: How are sorted the numerical features in order to find the
+      splits
       - PRESORT: The features are pre-sorted at the start of the training. This
         solution is faster but consumes much more memory than IN_NODE.
       - IN_NODE: The features are sorted just before being used in the node.
         This solution is slow but consumes little amount of memory.
       . Default: "PRESORT".
     sparse_oblique_normalization: For sparse oblique splits i.e.
-      `split_axis=SPARSE_OBLIQUE`. Normalization applied on the features,
-      before applying the sparse oblique projections.
+      `split_axis=SPARSE_OBLIQUE`. Normalization applied on the features, before
+      applying the sparse oblique projections.
       - `NONE`: No normalization.
       - `STANDARD_DEVIATION`: Normalize the feature by the estimated standard
         deviation on the entire train dataset. Also known as Z-Score
@@ -576,14 +810,14 @@ class GradientBoostedTreesModel(core.CoreModel):
     sparse_oblique_num_projections_exponent: For sparse oblique splits i.e.
       `split_axis=SPARSE_OBLIQUE`. Controls of the number of random projections
       to test at each node as `num_features^num_projections_exponent`. Default:
-      None.
+        None.
     sparse_oblique_projection_density_factor: For sparse oblique splits i.e.
       `split_axis=SPARSE_OBLIQUE`. Controls of the number of random projections
       to test at each node as `num_features^num_projections_exponent`. Default:
-      None.
+        None.
     split_axis: What structure of split to consider for numerical features.
-      - `AXIS_ALIGNED`: Axis aligned splits (i.e. one condition at a time).
-        This is the "classical" way to train a tree. Default value.
+      - `AXIS_ALIGNED`: Axis aligned splits (i.e. one condition at a time). This
+        is the "classical" way to train a tree. Default value.
       - `SPARSE_OBLIQUE`: Sparse oblique splits (i.e. splits one a small number
         of features) from "Sparse Projection Oblique Random Forests", Tomita et
         al., 2020. Default: "AXIS_ALIGNED".
@@ -592,9 +826,12 @@ class GradientBoostedTreesModel(core.CoreModel):
     use_hessian_gain: Use true, uses a formulation of split gain with a hessian
       term i.e. optimizes the splits to minimize the variance of "gradient /
       hessian. Available for all losses except regression. Default: False.
+    validation_interval_in_trees: Evaluate the model on the validation set every
+      "validation_interval_in_trees" trees. Increasing this value reduce the
+      cost of validation and can impact the early stopping policy (as early
+      stopping is only tested during the validation). Default: 1.
     validation_ratio: Ratio of the training dataset used to monitor the
       training. Require to be >0 if early stopping is enabled. Default: 0.1.
-
   """
 
   @core._list_explicit_arguments
@@ -612,6 +849,7 @@ class GradientBoostedTreesModel(core.CoreModel):
       advanced_arguments: Optional[AdvancedArguments] = None,
       num_threads: Optional[int] = 6,
       name: Optional[str] = None,
+      max_vocab_count: Optional[int] = 2000,
       adapt_subsample_for_maximum_training_duration: Optional[bool] = False,
       allow_na_conditions: Optional[bool] = False,
       apply_link_function: Optional[bool] = True,
@@ -619,6 +857,7 @@ class GradientBoostedTreesModel(core.CoreModel):
       categorical_set_split_greedy_sampling: Optional[float] = 0.1,
       categorical_set_split_max_num_items: Optional[int] = -1,
       categorical_set_split_min_item_frequency: Optional[int] = 1,
+      compute_permutation_variable_importance: Optional[bool] = False,
       dart_dropout: Optional[float] = 0.01,
       early_stopping: Optional[str] = "LOSS_INCREASE",
       early_stopping_num_trees_look_ahead: Optional[int] = 30,
@@ -627,18 +866,21 @@ class GradientBoostedTreesModel(core.CoreModel):
       goss_beta: Optional[float] = 0.1,
       growing_strategy: Optional[str] = "LOCAL",
       in_split_min_examples_check: Optional[bool] = True,
+      keep_non_leaf_label_distribution: Optional[bool] = True,
       l1_regularization: Optional[float] = 0.0,
       l2_categorical_regularization: Optional[float] = 1.0,
       l2_regularization: Optional[float] = 0.0,
       lambda_loss: Optional[float] = 1.0,
       max_depth: Optional[int] = 6,
       max_num_nodes: Optional[int] = None,
+      maximum_model_size_in_memory_in_bytes: Optional[float] = -1.0,
       maximum_training_duration_seconds: Optional[float] = -1.0,
       min_examples: Optional[int] = 5,
       missing_value_policy: Optional[str] = "GLOBAL_IMPUTATION",
       num_candidate_attributes: Optional[int] = -1,
       num_candidate_attributes_ratio: Optional[float] = -1.0,
       num_trees: Optional[int] = 300,
+      random_seed: Optional[int] = 123456,
       sampling_method: Optional[str] = "NONE",
       selective_gradient_boosting_ratio: Optional[float] = 0.01,
       shrinkage: Optional[float] = 0.1,
@@ -649,6 +891,7 @@ class GradientBoostedTreesModel(core.CoreModel):
       split_axis: Optional[str] = "AXIS_ALIGNED",
       subsample: Optional[float] = 1.0,
       use_hessian_gain: Optional[bool] = False,
+      validation_interval_in_trees: Optional[int] = 1,
       validation_ratio: Optional[float] = 0.1,
       explicit_args: Optional[Set[str]] = None):
 
@@ -667,6 +910,8 @@ class GradientBoostedTreesModel(core.CoreModel):
             categorical_set_split_max_num_items,
         "categorical_set_split_min_item_frequency":
             categorical_set_split_min_item_frequency,
+        "compute_permutation_variable_importance":
+            compute_permutation_variable_importance,
         "dart_dropout":
             dart_dropout,
         "early_stopping":
@@ -683,6 +928,8 @@ class GradientBoostedTreesModel(core.CoreModel):
             growing_strategy,
         "in_split_min_examples_check":
             in_split_min_examples_check,
+        "keep_non_leaf_label_distribution":
+            keep_non_leaf_label_distribution,
         "l1_regularization":
             l1_regularization,
         "l2_categorical_regularization":
@@ -695,6 +942,8 @@ class GradientBoostedTreesModel(core.CoreModel):
             max_depth,
         "max_num_nodes":
             max_num_nodes,
+        "maximum_model_size_in_memory_in_bytes":
+            maximum_model_size_in_memory_in_bytes,
         "maximum_training_duration_seconds":
             maximum_training_duration_seconds,
         "min_examples":
@@ -707,6 +956,8 @@ class GradientBoostedTreesModel(core.CoreModel):
             num_candidate_attributes_ratio,
         "num_trees":
             num_trees,
+        "random_seed":
+            random_seed,
         "sampling_method":
             sampling_method,
         "selective_gradient_boosting_ratio":
@@ -727,6 +978,8 @@ class GradientBoostedTreesModel(core.CoreModel):
             subsample,
         "use_hessian_gain":
             use_hessian_gain,
+        "validation_interval_in_trees":
+            validation_interval_in_trees,
         "validation_ratio":
             validation_ratio,
     }
@@ -749,7 +1002,8 @@ class GradientBoostedTreesModel(core.CoreModel):
         verbose=verbose,
         advanced_arguments=advanced_arguments,
         num_threads=num_threads,
-        name=name)
+        name=name,
+        max_vocab_count=max_vocab_count)
 
   @staticmethod
   def predefined_hyperparameters() -> List[core.HyperParameterTemplate]:
@@ -772,18 +1026,23 @@ class GradientBoostedTreesModel(core.CoreModel):
             }),
     ]
 
+  @staticmethod
+  def capabilities() -> abstract_learner_pb2.LearnerCapabilities:
+    return abstract_learner_pb2.LearnerCapabilities(
+        support_partial_cache_dataset_format=False)
+
 
 class RandomForestModel(core.CoreModel):
   r"""Random Forest learning algorithm.
 
   A Random Forest (https://www.stat.berkeley.edu/~breiman/randomforest2001.pdf)
   is a collection of deep CART decision trees trained independently and without
-  pruning. Each tree is trained on a random subset of the original training 
+  pruning. Each tree is trained on a random subset of the original training
   dataset (sampled with replacement).
-  
+
   The algorithm is unique in that it is robust to overfitting, even in extreme
   cases e.g. when there is more features than training examples.
-  
+
   It is probably the most well-known of the Decision Forest training
   algorithms.
 
@@ -821,9 +1080,9 @@ class RandomForestModel(core.CoreModel):
       models on top of each other. Unlike preprocessing done in the tf.dataset,
       the operation in "preprocessing" are serialized with the model.
     postprocessing: Like "preprocessing" but applied on the model output.
-    ranking_group: Only for `task=Task.RANKING`. Name of a tf.string feature that
-      identifies queries in a query/document ranking task. The ranking group
-      is not added automatically for the set of features if
+    ranking_group: Only for `task=Task.RANKING`. Name of a tf.string feature
+      that identifies queries in a query/document ranking task. The ranking
+      group is not added automatically for the set of features if
       `exclude_non_specified_features=false`.
     temp_directory: Temporary directory used to store the model Assets after the
       training, and possibly as a work directory during the training. This
@@ -836,19 +1095,18 @@ class RandomForestModel(core.CoreModel):
       If None (default) the default parameters of the library are used. If set,
       `default_hyperparameter_template` refers to one of the following
       preconfigured hyper-parameter sets. Those sets outperforms the default
-      hyper-parameters (either generally or in specific scenarios).
-      You can omit the version (e.g. remove "@v5") to use the last version of
-      the template. In this case, the hyper-parameter can change in between
-      releases (not recommended for training in production).
+      hyper-parameters (either generally or in specific scenarios). You can omit
+      the version (e.g. remove "@v5") to use the last version of the template.
+      In this case, the hyper-parameter can change in between releases (not
+      recommended for training in production).
       - better_default@v1: A configuration that is generally better than the
         default parameters without being more expensive. The parameters are:
-        winner_take_all=True.
+          winner_take_all=True.
       - benchmark_rank1@v1: Top ranking hyper-parameters on our benchmark
         slightly modified to run in reasonable time. The parameters are:
-        winner_take_all=True, categorical_algorithm="RANDOM",
-        split_axis="SPARSE_OBLIQUE", sparse_oblique_normalization="MIN_MAX",
-        sparse_oblique_num_projections_exponent=1.0.
-
+          winner_take_all=True, categorical_algorithm="RANDOM",
+          split_axis="SPARSE_OBLIQUE", sparse_oblique_normalization="MIN_MAX",
+          sparse_oblique_num_projections_exponent=1.0.
     advanced_arguments: Advanced control of the model that most users won't need
       to use. See `AdvancedArguments` for details.
     num_threads: Number of threads used to train the model. Different learning
@@ -856,10 +1114,15 @@ class RandomForestModel(core.CoreModel):
       efficiency. If specified, `num_threads` field of the
       `advanced_arguments.yggdrasil_deployment_config` has priority.
     name: The name of the model.
+    max_vocab_count: Default maximum size of the vocabulary for CATEGORICAL and
+      CATEGORICAL_SET features stored as strings. If more unique values exist,
+      only the most frequent values are kept, and the remaining values are
+      considered as out-of-vocabulary. The value `max_vocab_count` defined in a
+      `FeatureUsage` (if any) takes precedence.
     adapt_bootstrap_size_ratio_for_maximum_training_duration: Control how the
-      maximum training duration (if set) is applied. If false, the training
-      stop when the time is used. If true, adapts the size of the sampled
-      dataset used to train each tree such that `num_trees` will train within
+      maximum training duration (if set) is applied. If false, the training stop
+      when the time is used. If true, adapts the size of the sampled dataset
+      used to train each tree such that `num_trees` will train within
       `maximum_training_duration`. Has no effect if there is no maximum
       training duration specified. Default: False.
     allow_na_conditions: If true, the tree training evaluates conditions of the
@@ -868,9 +1131,8 @@ class RandomForestModel(core.CoreModel):
       - `CART`: CART algorithm. Find categorical splits of the form "value \\in
         mask". The solution is exact for binary classification, regression and
         ranking. It is approximated for multi-class classification. This is a
-        good first algorithm to use. In case of overfitting (very small
-        dataset, large dictionary), the "random" algorithm is a good
-        alternative.
+        good first algorithm to use. In case of overfitting (very small dataset,
+        large dictionary), the "random" algorithm is a good alternative.
       - `ONE_HOT`: One-hot encoding. Find the optimal categorical split of the
         form "attribute == param". This method is similar (but more efficient)
         than converting converting each possible categorical value into a
@@ -891,7 +1153,7 @@ class RandomForestModel(core.CoreModel):
       available, the least frequent items are ignored. Changing this value is
       similar to change the "max_vocab_count" before loading the dataset, with
       the following exception: With `max_vocab_count`, all the remaining items
-      are grouped in a special Out-of-vocabulary item. With `max_num_items`,
+        are grouped in a special Out-of-vocabulary item. With `max_num_items`,
       this is not the case. Default: -1.
     categorical_set_split_min_item_frequency: For categorical set splits e.g.
       texts. Minimum number of occurrences of an item to be considered.
@@ -907,31 +1169,41 @@ class RandomForestModel(core.CoreModel):
         words, as long as a node satisfy the splits "constraints (e.g. maximum
         depth, minimum number of observations), the node will be split. This is
         the "classical" way to grow decision trees.
-      - `BEST_FIRST_GLOBAL`: The node with the best loss reduction among all
-        the nodes of the tree is selected for splitting. This method is also
-        called "best first" or "leaf-wise growth". See "Best-first decision
+      - `BEST_FIRST_GLOBAL`: The node with the best loss reduction among all the
+        nodes of the tree is selected for splitting. This method is also called
+        "best first" or "leaf-wise growth". See "Best-first decision
         tree learning", Shi and "Additive logistic regression : A statistical
         view of boosting", Friedman for more details. Default: "LOCAL".
     in_split_min_examples_check: Whether to check the `min_examples` constraint
       in the split search (i.e. splits leading to one child having less than
-      `min_examples` examples are considered invalid) or before the split
-      search (i.e. a node can be derived only if it contains more than
-      `min_examples` examples). If false, there can be nodes with less than
+      `min_examples` examples are considered invalid) or before the split search
+      (i.e. a node can be derived only if it contains more than `min_examples`
+      examples). If false, there can be nodes with less than
       `min_examples` training examples. Default: True.
+    keep_non_leaf_label_distribution: Whether to keep the node value (i.e. the
+      distribution of the labels of the training examples) of non-leaf nodes.
+      This information is not used during serving, however it can be used for
+      model interpretation as well as hyper parameter tuning. In the worst
+      case, this can account for half of the model size. Default: True.
     max_depth: Maximum depth of the tree. `max_depth=1` means that all trees
       will be roots. Negative values are ignored. Default: 16.
     max_num_nodes: Maximum number of nodes in the tree. Set to -1 to disable
       this limit. Only available for `growing_strategy=BEST_FIRST_GLOBAL`.
       Default: None.
+    maximum_model_size_in_memory_in_bytes: Limit the size of the model when
+      stored in ram. Different algorithms can enforce this limit differently.
+      Note that when models are compiled into an inference, the size of the
+      inference engine is generally much smaller than the original model.
+      Default: -1.0.
     maximum_training_duration_seconds: Maximum training duration of the model
       expressed in seconds. Each learning algorithm is free to use this
       parameter at it sees fit. Enabling maximum training duration makes the
       model training non-deterministic. Default: -1.0.
     min_examples: Minimum number of examples in a node. Default: 5.
     missing_value_policy: Method used to handle missing attribute values.
-      - `GLOBAL_IMPUTATION`: Missing attribute values are imputed, with the
-        mean (in case of numerical attribute) or the most-frequent-item (in
-        case of categorical attribute) computed on the entire dataset (i.e. the
+      - `GLOBAL_IMPUTATION`: Missing attribute values are imputed, with the mean
+        (in case of numerical attribute) or the most-frequent-item (in case of
+        categorical attribute) computed on the entire dataset (i.e. the
         information contained in the data spec).
       - `LOCAL_IMPUTATION`: Missing attribute values are imputed with the mean
         (numerical attribute) or most-frequent-item (in the case of categorical
@@ -945,8 +1217,8 @@ class RandomForestModel(core.CoreModel):
       node. An attribute is valid if it has at least a valid split. If
       `num_candidate_attributes=0`, the value is set to the classical default
       value for Random Forest: `sqrt(number of input attributes)` in case of
-      classification and `number_of_input_attributes / 3` in case of
-      regression. If `num_candidate_attributes=-1`, all the attributes are
+        classification and `number_of_input_attributes / 3` in case of
+        regression. If `num_candidate_attributes=-1`, all the attributes are
       tested. Default: 0.
     num_candidate_attributes_ratio: Ratio of attributes tested at each node. If
       set, it is equivalent to `num_candidate_attributes =
@@ -956,16 +1228,18 @@ class RandomForestModel(core.CoreModel):
     num_trees: Number of individual decision trees. Increasing the number of
       trees can increase the quality of the model at the expense of size,
       training speed, and inference latency. Default: 300.
-    sorting_strategy: How are sorted the numerical features in order to find
-      the splits
+    random_seed: Random seed for the training of the model. Learners are
+      expected to be deterministic by the random seed. Default: 123456.
+    sorting_strategy: How are sorted the numerical features in order to find the
+      splits
       - PRESORT: The features are pre-sorted at the start of the training. This
         solution is faster but consumes much more memory than IN_NODE.
       - IN_NODE: The features are sorted just before being used in the node.
         This solution is slow but consumes little amount of memory.
       . Default: "PRESORT".
     sparse_oblique_normalization: For sparse oblique splits i.e.
-      `split_axis=SPARSE_OBLIQUE`. Normalization applied on the features,
-      before applying the sparse oblique projections.
+      `split_axis=SPARSE_OBLIQUE`. Normalization applied on the features, before
+      applying the sparse oblique projections.
       - `NONE`: No normalization.
       - `STANDARD_DEVIATION`: Normalize the feature by the estimated standard
         deviation on the entire train dataset. Also known as Z-Score
@@ -975,22 +1249,21 @@ class RandomForestModel(core.CoreModel):
     sparse_oblique_num_projections_exponent: For sparse oblique splits i.e.
       `split_axis=SPARSE_OBLIQUE`. Controls of the number of random projections
       to test at each node as `num_features^num_projections_exponent`. Default:
-      None.
+        None.
     sparse_oblique_projection_density_factor: For sparse oblique splits i.e.
       `split_axis=SPARSE_OBLIQUE`. Controls of the number of random projections
       to test at each node as `num_features^num_projections_exponent`. Default:
-      None.
+        None.
     split_axis: What structure of split to consider for numerical features.
-      - `AXIS_ALIGNED`: Axis aligned splits (i.e. one condition at a time).
-        This is the "classical" way to train a tree. Default value.
+      - `AXIS_ALIGNED`: Axis aligned splits (i.e. one condition at a time). This
+        is the "classical" way to train a tree. Default value.
       - `SPARSE_OBLIQUE`: Sparse oblique splits (i.e. splits one a small number
         of features) from "Sparse Projection Oblique Random Forests", Tomita et
         al., 2020. Default: "AXIS_ALIGNED".
     winner_take_all: Control how classification trees vote. If true, each tree
       votes for one class. If false, each tree vote for a distribution of
       classes. winner_take_all_inference=false is often preferable. Default:
-      True.
-
+        True.
   """
 
   @core._list_explicit_arguments
@@ -1008,6 +1281,7 @@ class RandomForestModel(core.CoreModel):
       advanced_arguments: Optional[AdvancedArguments] = None,
       num_threads: Optional[int] = 6,
       name: Optional[str] = None,
+      max_vocab_count: Optional[int] = 2000,
       adapt_bootstrap_size_ratio_for_maximum_training_duration: Optional[
           bool] = False,
       allow_na_conditions: Optional[bool] = False,
@@ -1019,14 +1293,17 @@ class RandomForestModel(core.CoreModel):
       compute_oob_variable_importances: Optional[bool] = False,
       growing_strategy: Optional[str] = "LOCAL",
       in_split_min_examples_check: Optional[bool] = True,
+      keep_non_leaf_label_distribution: Optional[bool] = True,
       max_depth: Optional[int] = 16,
       max_num_nodes: Optional[int] = None,
+      maximum_model_size_in_memory_in_bytes: Optional[float] = -1.0,
       maximum_training_duration_seconds: Optional[float] = -1.0,
       min_examples: Optional[int] = 5,
       missing_value_policy: Optional[str] = "GLOBAL_IMPUTATION",
       num_candidate_attributes: Optional[int] = 0,
       num_candidate_attributes_ratio: Optional[float] = -1.0,
       num_trees: Optional[int] = 300,
+      random_seed: Optional[int] = 123456,
       sorting_strategy: Optional[str] = "PRESORT",
       sparse_oblique_normalization: Optional[str] = None,
       sparse_oblique_num_projections_exponent: Optional[float] = None,
@@ -1056,10 +1333,14 @@ class RandomForestModel(core.CoreModel):
             growing_strategy,
         "in_split_min_examples_check":
             in_split_min_examples_check,
+        "keep_non_leaf_label_distribution":
+            keep_non_leaf_label_distribution,
         "max_depth":
             max_depth,
         "max_num_nodes":
             max_num_nodes,
+        "maximum_model_size_in_memory_in_bytes":
+            maximum_model_size_in_memory_in_bytes,
         "maximum_training_duration_seconds":
             maximum_training_duration_seconds,
         "min_examples":
@@ -1072,6 +1353,8 @@ class RandomForestModel(core.CoreModel):
             num_candidate_attributes_ratio,
         "num_trees":
             num_trees,
+        "random_seed":
+            random_seed,
         "sorting_strategy":
             sorting_strategy,
         "sparse_oblique_normalization":
@@ -1104,7 +1387,8 @@ class RandomForestModel(core.CoreModel):
         verbose=verbose,
         advanced_arguments=advanced_arguments,
         num_threads=num_threads,
-        name=name)
+        name=name,
+        max_vocab_count=max_vocab_count)
 
   @staticmethod
   def predefined_hyperparameters() -> List[core.HyperParameterTemplate]:
@@ -1126,3 +1410,8 @@ class RandomForestModel(core.CoreModel):
                 "sparse_oblique_num_projections_exponent": 1.0
             }),
     ]
+
+  @staticmethod
+  def capabilities() -> abstract_learner_pb2.LearnerCapabilities:
+    return abstract_learner_pb2.LearnerCapabilities(
+        support_partial_cache_dataset_format=False)